Allow compilation environment variables to override the platform string.
The `dart.os.name` and `dart.os.version` compilation environment variables,
if set, completely overrides the values of the global
`operatingSystem` and `operatingSystemVersion` getters.

By having a constant first priority, setting those values may allow compilers
to recognize the value as being a constant, and therefore tree-shake more
code. The value should be set to something that actually *works*
for the platform being compiled to.