Use the VM's "platform-const" feature to achieve tree-shaking. (#31)
* Use the VM's "platform-const" feature to achieve tree-shaking.

By using the VM's `@pragma("vm:platform-const")` annotation,
and `Platform` members annotated with it, ensure that the
`isLinux` and `operatingSystem` values are also known at compile-time,
even if they cannot be annotated as platform-constants.

This is done by having a class per known operating system,
and making sure to only create instances of those classes
guarded by platform-constants (or real constants), so that
if nobody goes out of their way to create a custom instance,
which they should only do for testing, all but one of those
platform-specific classes can be tree-shaken.

At that point, checks like `is MacOS` (one of the types)
can be statically determined to be false if `MacOS` was tree-shaken,
and `.id` can be statically resolved to a single final variable
with a known value.

* Update test matrix to start at 3.0.0

* Add example discussing tree-shaking.

* Format

* Address comments.

* Don't refer to unimported name in docs.